requirements.txt
alembic==1.3.1
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
blinker==1.4
Click==7.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
importlib-metadata==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
lazy-object-proxy==1.4.3
Mako==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mirakuru==2.1.0
mock==3.0.5
more-itertools==8.0.0
packaging==19.2
pluggy==0.13.1
port-for==0.4
psutil==5.6.7
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.1.0
pylint-plugin-utils==0.6
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.0
pytz==2019.3
rope==0.14.0
six==1.13.0
SQLAlchemy==1.3.11
typed-ast==1.4.0
urllib3==1.25.7
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
